--- a/kolibri/core/auth/serializers.py
+++ b/kolibri/core/auth/serializers.py
@@ -44,6 +44,7 @@ class FacilityUserSerializer(serializers.ModelSerializer):
             "gender",
             "birth_year",
         )
+        read_only_fields = ("is_superuser",)
 
     def validate(self, attrs):
         username = attrs.get("username")