… kolibri/core/auth/management/commands/fullfacilitysync.py
… kolibri/core/auth/management/utils.py
… kolibri/core/auth/models.py
… kolibri/core/auth/test/helpers.py
… kolibri/core/auth/test/test_api.py
… kolibri/core/content/api.py
… kolibri/core/content/models.py
… kolibri/core/content/serializers.py
… kolibri/core/content/utils/paths.py
… kolibri/core/device/management/commands/provisiondevice.py
… kolibri/core/device/models.py
… kolibri/core/device/serializers.py
… kolibri/core/device/translation.py
… kolibri/core/device/utils.py
… kolibri/core/lessons/serializers.py
… kolibri/core/lessons/viewsets.py
… kolibri/core/query.py
… kolibri/core/templates/kolibri/base.html
… kolibri/deployment/default/dev_urls.py
… kolibri/plugins/coach/class_summary_api.py
… kolibri/plugins/learn/serializers.py
… kolibri/plugins/learn/viewsets.py
… kolibri/utils/api.py