### Summary
* Adds `ValuesViewset` that allows continued use of DRF Serializers for write operations (to help with validation etc), but allows for limiting the overall number of queries required for read operations by using `values` calls instead of resolving a queryset and then doing subsequent calls on the model objects.
* Applies `ValuesViewset` to `ContentNodeSlimViewset` - reducing queries to retrieve a list of contentnodes.
* Applies `ValuesViewset` to `ClassroomViewset` - further optimizes the classroom viewset, reducing from ~100 queries for 26 classrooms to ~4 queries.
* Applies `ValuesViewset` to `FacilityUserViewset` - reducing queries from ~85 queries for ~40 users to ~2 queries.
* Applies `ValuesViewset` to `FacilityUserNameViewset` to avoid the need to use a serializer at all.
* Applies `ValuesViewset` to `LearnerGroupViewset` - reducing queries from N+4 queries to 4 queries.
* Removes use of DRF serializers from the classroom summary API endpint - reducing queries for a class with multiple quizzes and lessons (from real user data) from 985 queries (with ~16s of CPU time for the request) to 18 queries (with ~1.5s of CPU time for the request).
* Updates `learn` `LearnerClassroomViewset` - reducing queries for a user in several classes with a few active exams and lessons from 31 queries (with ~2700ms of CPU time for the request) to 13 queries (with ~400ms of CPU time for the request) - perhaps most dramatically, changes during this optimization reduces the time for the query on my SSD from ~700ms to <10ms, mostly by removing use of the `HierarchicalRelationsFilter` - it appears to interact particularly badly when used in querysets that are used as subqueries.
* Consolidates logic for device settings, and caches all device settings calls in common. Loading any page `facility`/`user` etc - will now result in no queries to the DB when the cache is hot.
* Removes the API schema generation logic, which was not frequently used, and resulted in errors sometimes.

### Reviewer guidance
Does the topics interface still function appropriately?
Does changing language still work?
Do the classroom management and facility user management parts of `facility` work properly?

### References
Hopefully fixes #6008 

----

### Contributor Checklist


PR process:

- [x] PR has the correct target branch and milestone
- [x] PR has 'needs review' or 'work-in-progress' label
- [x] If PR is ready for review, a reviewer has been added. (Don't use 'Assignees')
- [ ] If this is an important user-facing change, PR or related issue has a 'changelog' label
- [ ] If this includes an internal dependency change, a link to the diff is provided

Testing:

- [x] Contributor has fully tested the PR manually
- [ ] If there are any front-end changes, before/after screenshots are included
- [ ] Critical user journeys are covered by Gherkin stories
- [x] Critical and brittle code paths are covered by unit tests

### Reviewer Checklist

- Automated test coverage is satisfactory
- PR is fully functional
- PR has been tested for [accessibility regressions](http://kolibri-dev.readthedocs.io/en/develop/manual_testing.html#accessibility-a11y-testing)
- External dependency files were updated if necessary (`yarn` and `pip`)
- Documentation is updated
- Contributor is in AUTHORS.md